@@ -1,10 +1,10 @@
 name: juxtapose
 description: A widget for comparing two stacked widgets by dragging a slider thumb to reveal either sides of the slider, either horizontally or vertically.
-version: 1.0.3
+version: 1.1.0-nullsafety.1
 homepage: https://github.com/lesliearkorful/juxtapose
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
